добавил вызов функции  creatingPhotoElement в main.js
