+++ js/picture-parameters.js
@@ -1,0 +1,9 @@
+const parameters = {
+  containerClass: '.pictures',
+  templateId: '#picture',
+  templateClass: '.picture',
+  urlClass: '.picture__img',
+  likesClass: '.picture__likes',
+  commentsClass: '.picture__comments',
+};
+export {parameters};
